Modify PowerController long press command to use duration (and large timeout). PowerMenuTest (b/383004544) was failing on tablets.
The long press event timeout threshold (i.e. time between pressing key
down and key up required to trigger long press) is 500 ms on phones but 750 ms on tablets.

With the new change, the duration triggered with the --longpress tag is
non-device specific and under the required threshold on tablets to
trigger long press, causing the test to fail.

Thus, we change the long press command to use duration instead with the
timeout specified by the config.

ABTD results of tests passing with this change:
https://android-build.corp.google.com/abtd/run/L66200030008426443/
https://android-build.corp.google.com/abtd/run/L94300030008415110/

Bug: 383004544
Test: atest PowerMenuTest (on bluejay/felix)
Test: atest ContextualTipsTest (on bluejay/felix)
Change-Id: I83c088507601c8f4dd4f34c3aac272a0dddfa4b2