Mitigate CF devices mistakenly interpreting a swipe as a long press
Long pressing the status bar now expands the shade (for a11y
compliance) but Cuttlefish devices have sever performance issues that
often cause them to misinterpret swipes as long presses. This CL adds
a utility for extending the timeout of the long press during a test
and restoring it to the original timeout when the test is complete.

Test: verified that real devices do not have this issue
Test: verified that the test fails without this mitigation
Test: verified that the test passes with this mitigation
Bug: 371224114
Flag: com.android.systemui.shade_expands_on_status_bar_long_press
Change-Id: I21414a53d4c0345b657b8c9be562db085ae19c59