Modify app increases in size assertion method.
The captured app layers during the test captures the initial task region, the resize veil and the transition. The value of the layers' region may not be in sequence, so filter by the resize veil layers.

Also allow a small tolerance in `coversAtLeast` to eliminate flakiness during the captured resize transition layers.

Fixes: 374727033
Fixes: 373638597
Flag: EXEMPT adding tests
Test: atest ResizeAppWithEdgeResizeMouse,
ResizeAppWithEdgeResizeTouchpad, ResizeAppWithEdgeResizeStylus,
EnterDesktopWithDragPortrait, ResizeAppWithCornerResizeLandscape

Change-Id: I0d832d7b6eb73979bd0e775257c32b8d79bec3da