Handle display scaling in DesktopMouseTestRule
- Added new LogicalPhysicalDisplayTransformHelper
- Updated DesktopMouseTestRule to factor in display scale
in both movement and verification
- Renamed bunch of Delta and Point class to a defined Physical/Logical
PX class

As discussed in b/444082242#comment2, some tests modify the display
scale which change how input is processed on native side. For example,
if display scale is 2.0f (logical display set to be twice of physical
display size), then sending 5px move to evdev will result to 10px
dispatched to WM, which made cursor to be inaccurate and caused failing
test

To fix this, a inverse display transform should be applied before
sending mouse events, the detail is abstracted in
LogicalPhysicalDisplayTransformHelper.

DesktopMouseTestRule should also consider scale when verifying final
position, because if display scale is 3.0f for example, the minimal
movement would be 3px on logical display (or 1px on physical display),
and hence the final position accuracy will be within 3px margin of
error.

For question why we can't send 0.3px of physical display (so that it
will result to 1px on logical display), it's because evdev REL X/Y only
receives integer delta, and anything below that will be no-op.

Therefore, DesktopMouseTestRule also provides a getMinMouseMovement
method for other tests to use, if they want to ensure that a move
event at least occurs

Bug: 444082242
Test: adb shell wm size 504x4488, rotate, atest MouseToTouchTest
Flag: EXEMPT test rule fix

Change-Id: Id07b65e0c7047c2e6bb5e766b7e5ec0e7bd6bd94