Add a color inversion test
drmhwc does not support color inverstion. This test verifies that when
color inversion is enabled, no change to the CTM is made.

Remove stop/start for teardown as these settings do not require it to
take effect. Move expensive device operations to before class init
method. Add a waiting period for the night light animatino to complete.

Bug: 444038726
Test: atest HardwareColorTests
Flag: Test Only
Change-Id: I57256e19ca6597b05d2f2e24d41e32be0e1d5742