Add Gemini API utility for UI element verification.
This change introduces a new Python library `ai_utils` under `//depot/libraries/automotive`. The library includes `gemini_util.py`, which provides a `GeminiUtil` class to interact with Gemini APIs. This utility is designed to analyze screenshots and verify the presence and details of specific UI elements, returning results in a structured JSON format.

Test: Tested Using CATBox and Atest
Bug:445934546
Change-Id: I1c080fa6b37378984ebbbed80289a5884b3d8fcd