Redesign Spectatio Host-Side Test Framework.
This change introduces a `TestContext` as a service locator to manage test dependencies, including devices and test arguments. The `TestDeviceManager` is created to handle device discovery and ordering. Custom exceptions are added for better error handling. The `TestDevice` class is also refactored for improved readability and type hinting.

ABTD Test Execution -
 * Screenshot Test - https://android-build.corp.google.com/test_investigate/invocation/I04700010457242805/test/TR68729927220300971/
 * Log Verification Test - https://android-build.corp.google.com/test_investigate/invocation/I85600010457237868/test/TR22829927223702743

Test: Locally using ATest and CATBox, ABTD
Bug: 445738759
Change-Id: Iaa7627b71b6f31b85089deb13302a3547f281014