Capture StatusBar icon positions by reusing UiObject2
Reuse the same UiObject2/HomeAppIcon objects
when capturing status bar icon positions.

When querying the objects we wait for the animation to finish
which makes the test fail as we expect to capture icon positions
midway through the animation.

We use the same approach in Launcher test, updated the
code here as well to match it.

Test: atest StatusBarFoldAnimationTest#testFold_viewsMovingTowardsCenter
Bug: 449303499
Flag: EXEMPT TEST_ONLY
Change-Id: I6914f36de9588959fa11424faa656389d86501fb