igt: Increase test timeout to 1hr
When IGT runs on VKMS with 3 displays at 4K, it takes a while.
As all the processing is done on CPU, some plane color format
(like YUV) will be very slow to render and create CRC timeouts.

kms_plane@pixel format took 48 mins, so let's increase the test timeout
to give it a chance to pass.

BUG: b/450275771
Test: atest IgtKmsPlaneTestCases
Change-Id: Ib2757f47854595a1bf53bc48d415c36d85061ac8