Add a check to ensure bubble bar is expanded
When creating a new ExpandedBubbleBar TAPL object, make sure the
expanded view exists. We can't have expanded bubble bar without the
expanded view.

Some tests rely on Root.get().bubbleBar.expand() to expand the bubble
bar and expect the expanded view to be there after. But we were not
checking for the view.

Make sure we have this check in place.

Bug: 450383432
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.bubble.BubbleBarTest
Flag: EXEMPT TEST_ONLY
Change-Id: Ib790b3a7ea6deb096424bbde380582fc5d65b4ff