Include removing all desks in rule
RemoveAllTasksButHomeRule removes recent tasks but leaves desks in
recents alive. Doing so may cause the |withHomeActivityVisible|
assertion to fail because in desktop-first displays, removing all tasks
in a desk will still leave the desk active and thus the desktop
wallpaper task visible instead of home.

Removing desks fixes this and is closer to the original intent of
RemoveAllTasksButHomeRule.

Flag: EXEMPT bugfix
Bug: 452066935
Test: atest FunctionalTestsDesktopMode
Change-Id: I3618b5f3b7387b70c4f7fe957be3acb387147af1