Add option to enable CUJ tracing on Perfetto monitor
The PerfettoTraceMonitor is updated to ensure that Critical User
Journeys (CUJs) are reliably captured in Perfetto traces. This is
accomplished by programmatically enabling the `interaction_jank_monitor`
and setting its sampling interval to 1 at the start of a tracing
session. The original values are restored when the session is stopped.

Test: atest FlickerLibUtilsTest:android.tools.monitors.PerfettoTraceMonitorTest#cujTracingTest --iterations 10
Fixes: 443270014
Change-Id: I22a8d0b29762d90f77da90a6927d717062c60dbb