Skip tests if the flag is optimized
This change will skip tests if the flag is optimized. A flag is
considered as optimized, if the Filed of FeatureFlags cannot be found in
the Flags class. For this case this tool cannot switch the underlying
implementation. In such case, this tool will check whether the test
tries to set the flag to its default value. If not, then skip the test.
The reason for not failing the test if the value on set is not equal to
the default value is it will break existing tests.

Test: atest SystemUITests
Bug: 447177427
Change-Id: I26fbd80c93f3f2b1b8206f3b86ad99ef9e09d3e6