Convert uints from unsigned string value from args table.
Perfetto are making changes to show uint values in the display_value
column as actual uints instead of converting them to signed ints. We
need to accordingly convert from string to ULong, then to Long, to
prevent having to deal with ULongs everywhere else in the flicker code.

Bug: 463983064
Test: atest FlickerLibTest
Flag: EXEMPT TEST FIX
Change-Id: I6934387e8d263f6f02e33449f3cda50882b93fb3