Reuse UiAutomation instance when adopting permissions
Fetching UiAutomation caused a11y services to be interrupted causing
magnifying service to fail. An alternative to this approach would be to
fetch UiAutomation with
instrumentation.getUiAutomation(FLAG_DONT_SUPPRESS_ACCESSIBILITY_SERVICES)
But as a test lib, it would be best to use UiAutomation without any
additional param, and to keep UiAutomation static as tests might be
relying on other UiAutomation running states as well.

Bug: 465498285
Test: atest FullScreenMagnificationMouseFollowingTest
Flag: EXEMPT BUGFIX
Change-Id: I29a690a01bba4ba7bb3eb5971143f76f893b1eef