feat(atf x ui automator): add API to disable checks, including for certain tests
If you add the rule to a base class that is used in hundreds of tests, you may not want to run the checks on a subset of those tests (for instance, performance tests). This provides a way to check the test class name and test method name, and disable all checks dynamically before each test.

Fix: 469461817

Flag: EXEMPT TEST_ONLY
Test: in google3

Change-Id: Ie5f29b9caba3107eb8913dfd18819e3910f173e6